requirements.txt
python-telegram-bot[job-queue]==20.7
fastapi==0.109.2
uvicorn[standard]==0.27.1
sqlalchemy[asyncio]==2.0.25
aiosqlite==0.19.0
aiomysql==0.2.0
alembic==1.13.1
httpx~=0.25.2
beautifulsoup4==4.12.3
bcrypt==4.1.2
pydantic==2.6.0
pydantic-settings==2.1.0
python-dotenv==1.0.1
asyncpg==0.29.0
lxml==5.1.0
